fix: 레이어드 아키텍쳐로 수정
