+++ src/main/java/com/petmatz/api/user/controller/UserController.java
@@ -1,0 +1,109 @@
+package com.petmatz.api.user.controller;
+
+import com.petmatz.user.request.*;
+import com.petmatz.user.response.*;
+import com.petmatz.domain.user.service.UserService;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@Slf4j
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/auth")
+public class UserController {
+    private final UserService userService;
+
+    @PostMapping("/email-certification")
+    public ResponseEntity<? super EmailCertificationResponseDto> emailCertification(@RequestBody @Valid EmailCertificationRequestDto requestBody) {
+        ResponseEntity<? super EmailCertificationResponseDto> response= userService.emailCertification(requestBody);
+        log.info("[emailCertification]: { accountId: " + requestBody.getAccountId() + "}");
+        return response;
+    }
+
+    @PostMapping("/check-certification")
+    public ResponseEntity<? super CheckCertificationResponseDto> checkCertification(@RequestBody @Valid CheckCertificationRequestDto requestBody) {
+        ResponseEntity<? super CheckCertificationResponseDto> response= userService.checkCertification(requestBody);
+        log.info("[checkCertification]: {accountId: " + requestBody.getAccountId() + ", certificationNumber: " + requestBody.getCertificationNumber() + "}");
+        return response;
+    }
+
+    @PostMapping("/sign-up")
+    public ResponseEntity<? super SignUpResponseDto> signUp(@RequestBody @Valid SignUpRequestDto requestBody) {
+        ResponseEntity<? super SignUpResponseDto> response= userService.signUp(requestBody);
+        log.info("[signUp]: { accountId: " + requestBody.getAccountId() + ", password: " + requestBody.getPassword());
+        return response;
+    }
+
+    @PostMapping("/sign-in")
+    public ResponseEntity<? super SignInResponseDto> signIn(
+            @RequestBody @Valid SignInRequestDto requestBody,
+            HttpServletResponse response) {
+
+        ResponseEntity<? super SignInResponseDto> result = userService.signIn(requestBody, response);
+        log.info("[signIn]: { accountId: " + requestBody.getAccountId() + ", result: " + result.getStatusCode() + " }");
+        return result;
+    }
+
+    @PostMapping("/delete-user")
+    public ResponseEntity<? super DeleteIdResponseDto> deleteUser(@RequestBody @Valid DeleteIdRequestDto requestBody) {
+        ResponseEntity<? super DeleteIdResponseDto> response= userService.deleteId(requestBody);
+        log.info("[deleteUser]:{password: " + requestBody.getPassword() + "}");
+        return response;
+    }
+
+    //---------------------------------------------------------------------------------------------------------------------------------//
+    @GetMapping("/get-myprofile")
+    public ResponseEntity<? super GetMyProfileResponseDto> getMypage() {
+        ResponseEntity<? super GetMyProfileResponseDto> response = userService.getMypage();
+        log.info("[getMypage]");
+        return response;
+    }
+
+    @GetMapping("/get-otherprofile")
+    public ResponseEntity<? super GetMyProfileResponseDto> getOtherMypage(@ModelAttribute @Valid GetMyProfileRequestDto requestBody) {
+        ResponseEntity<? super GetMyProfileResponseDto> response = userService.getOtherMypage(requestBody);
+        log.info("[getOtherMypage]");
+        return response;
+    }
+
+    @PostMapping("/edit-myprofile")
+    public ResponseEntity<? super EditMyProfileResponseDto> editMyProfile(@RequestBody @Valid EditMyProfileRequestDto requestBody) {
+        ResponseEntity<? super EditMyProfileResponseDto> response = userService.editMyProfile(requestBody);
+        log.info("[editMyProfile]");
+        return response;
+    }
+
+    @PostMapping("/send-repassword")
+    public ResponseEntity<? super SendRepasswordResponseDto> sendRepassword(@RequestBody @Valid SendRepasswordRequestDto requestBody) {
+        ResponseEntity<? super SendRepasswordResponseDto> response = userService.sendRepassword(requestBody);
+        log.info("[sendRepassword]: {accountId: " + requestBody.getAccountId()+"}");
+        return response;
+    }
+
+    @PostMapping("/repassword")
+    public ResponseEntity<? super RepasswordResponseDto> repassword(@RequestBody @Valid RepasswordRequestDto requestBody) {
+        ResponseEntity<? super RepasswordResponseDto> response = userService.repassword(requestBody);
+        log.info("[repassword]: {currentPassword: " + requestBody.getCurrentPassword() + ", newPassword: " + requestBody.getNewPassword() + "}");
+        return response;
+    }
+    //---------------------------------------------------------------------------------------------------------------------------------------------------------------
+    @PostMapping("/hearting")
+    public ResponseEntity<? super HeartingResponseDto> hearting(@RequestBody @Valid HeartingRequestDto requestBody) {
+        ResponseEntity<? super HeartingResponseDto> response = userService.hearting(requestBody);
+        log.info("[hearting]: {heartedId: " + requestBody.getHeartedId() +"}");
+        return response;
+    }
+
+    @GetMapping("/get-heartlist")
+    public ResponseEntity<? super GetHeartingListResponseDto>getHeartedList() {
+        ResponseEntity<? super GetHeartingListResponseDto> response = userService.getHeartedList();
+        log.info("[getHeartedList]");
+        return response;
+    }
+
+}
+
