+++ src/main/java/com/petmatz/common/security/handler/ValidationExceptionHandler.java
@@ -1,0 +1,16 @@
+package com.petmatz.common.security.handler;
+
+import com.petmatz.user.common.LogInResponseDto;
+import org.springframework.http.ResponseEntity;
+import org.springframework.http.converter.HttpMessageNotReadableException;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class ValidationExceptionHandler {
+    @ExceptionHandler({MethodArgumentNotValidException.class, HttpMessageNotReadableException.class})
+    public ResponseEntity<LogInResponseDto> validationExceptionHandler(Exception e) {
+        return LogInResponseDto.validationFail();
+    }
+}
