+++ src/main/java/com/petmatz/domain/user/entity/Heart.java
@@ -1,0 +1,28 @@
+package com.petmatz.domain.user.entity;
+
+import com.petmatz.domain.global.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+
+@Getter
+@SuperBuilder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@Entity(name = "Heart")
+@Table(name = "Heart")
+public class Heart extends BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    @Column(name = "my_id", nullable = false)
+    private Long myId;
+
+    @Column(name = "hearted_id", nullable = false)
+    private Long heartedId;
+}
