+++ src/main/java/com/petmatz/domain/user/service/EmailProvider.java
@@ -1,0 +1,5 @@
+package com.petmatz.domain.user.service;
+
+public interface EmailProvider {
+    boolean sendVerificationEmail(String email, String certificationNumber);
+}
