+++ src/main/java/com/petmatz/domain/user/service/RePasswordEmailProvider.java
@@ -1,0 +1,5 @@
+package com.petmatz.domain.user.service;
+
+public interface RePasswordEmailProvider {
+    boolean sendVerificationEmail(String email, String rePassword);
+}
