--- src/main/java/com/petmatz/user/entity/Certification.java
@@ -1,31 +1,0 @@
-package com.petmatz.user.entity;
-
-import com.petmatz.domain.global.BaseEntity;
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@Getter
-@SuperBuilder
-@NoArgsConstructor(access= AccessLevel.PROTECTED)
-@AllArgsConstructor
-@Entity(name="Certification" )
-@Table(name = "Certification"  )
-/**
- * 이메일 검증을 위한 Entity
- */
-public class Certification extends BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", nullable = false)
-    private Long id;
-
-    @Column(name = "account_id", nullable = false, length = 30)
-    private String accountId;
-
-    @Column(name = "certification_number", nullable = false, length = 6)
-    private String certificationNumber;
-}
