--- src/main/java/com/petmatz/user/provider/RePasswordProvider.java
@@ -1,15 +1,0 @@
-package com.petmatz.user.provider;
-
-import java.util.UUID;
-
-public class RePasswordProvider {
-    public static String generatePassword() {
-        // UUID 생성 후 하이픈을 제거한 문자열로 변환
-        String uuid = UUID.randomUUID().toString().replaceAll("-", "");
-
-        // UUID에서 영문자와 숫자로만 이루어진 10자리 문자열 추출
-        String number = uuid.substring(0, 10);
-
-        return number;
-    }
-}
