--- src/main/java/com/petmatz/user/repository/HeartRepository.java
@@ -1,14 +1,0 @@
-package com.petmatz.user.repository;
-
-import com.petmatz.user.entity.Heart;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface HeartRepository extends JpaRepository<Heart, String> {
-    boolean existsByMyIdAndHeartedId(Long myId, Long heartedId);
-    void deleteByMyIdAndHeartedId(Long myId, Long heartedId);
-
-    void deleteAllByHeartedId(Long heartedId);
-    List<Heart> findAllByMyId(Long myId);
-}
