--- src/main/java/com/petmatz/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.petmatz.user.repository;
-
-import com.petmatz.user.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, String> {
-    boolean existsById(Long userId);
-    boolean existsByAccountId(String accountId);
-    User findById(Long userId);
-    User findByAccountId(String accountId);
-}
