+++ src/main/java/com/petmatz/user/request/CheckCertificationRequestDto.java
@@ -1,5 +1,6 @@
 package com.petmatz.user.request;
 
+import com.petmatz.user.request.info.CheckCertificationInfo;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -14,4 +15,11 @@ public class CheckCertificationRequestDto {
 
     @NotBlank
     private String certificationNumber;
+
+    public static CheckCertificationInfo of(CheckCertificationRequestDto reqDto) {
+        return CheckCertificationInfo.builder()
+                .accountId(reqDto.getAccountId())
+                .certificationNumber(reqDto.getCertificationNumber())
+                .build();
+    }
 }
