+++ src/main/java/com/petmatz/user/request/info/CheckCertificationInfo.java
@@ -1,0 +1,12 @@
+package com.petmatz.user.request.info;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class CheckCertificationInfo {
+    private String accountId;
+
+    private String certificationNumber;
+}
