+++ src/main/java/com/petmatz/user/request/info/EditMyProfileInfo.java
@@ -1,0 +1,20 @@
+package com.petmatz.user.request.info;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class EditMyProfileInfo {
+    private String nickname;
+
+    private String preferredSize;
+
+    private String introduction;
+
+    private boolean isCareAvailable;
+
+    private Integer timeWage;
+
+    private Integer monthWage;
+}
