+++ src/main/java/com/petmatz/user/request/info/RepasswordInfo.java
@@ -1,0 +1,11 @@
+package com.petmatz.user.request.info;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class RepasswordInfo {
+    private String currentPassword;
+    private String newPassword;
+}
