+++ src/main/java/com/petmatz/user/request/info/SignInInfo.java
@@ -1,0 +1,11 @@
+package com.petmatz.user.request.info;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class SignInInfo {
+    private String accountId;
+    private String password;
+}
