+++ src/main/java/com/petmatz/user/request/info/SignUpInfo.java
@@ -1,0 +1,31 @@
+package com.petmatz.user.request.info;
+
+import com.petmatz.domain.user.entity.User;
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class SignUpInfo {
+    private String accountId;
+
+    private String password;
+
+    private String certificationNumber;
+
+    private String nickname;
+
+    private User.Gender gender;
+
+    private String preferredSize;
+
+    private Boolean isCareAvailable;
+
+    private Integer timeWage;
+
+    private Integer monthWage;
+
+    private String introduction;
+
+    private String mbti;
+}
