+++ src/main/java/com/petmatz/user/response/GetHeartingListResponseDto.java
@@ -1,7 +1,7 @@
 package com.petmatz.user.response;
 
 import com.petmatz.user.common.LogInResponseDto;
-import com.petmatz.user.entity.Heart;
+import com.petmatz.domain.user.entity.Heart;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
