+++ src/main/java/com/petmatz/user/response/GetMyProfileResponseDto.java
@@ -3,7 +3,7 @@ package com.petmatz.user.response;
 import com.petmatz.user.common.LogInResponseDto;
 import com.petmatz.user.common.ResponseCode;
 import com.petmatz.user.common.ResponseMessage;
-import com.petmatz.user.entity.User;
+import com.petmatz.domain.user.entity.User;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
